# api/manager/swagger.yaml
basePath: /
definitions:
...
  d7y_io_dragonfly_v2_manager_models.Config:
    properties:
      bio:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_del:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.User'
      user_id:
        type: integer
      value:
        type: string
    type: object
...
  d7y_io_dragonfly_v2_manager_models.User:
    properties:
      avatar:
        type: string
      bio:
        type: string
      configs:
        items:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.Config'
        type: array
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_del:
        type: integer
      location:
        type: string
      name:
        type: string
      phone:
        type: string
      state:
        type: string
      updated_at:
        type: string
    type: object
...
  d7y_io_dragonfly_v2_manager_types.ResetPasswordRequest:
    properties:
      new_password:
        maxLength: 20
        minLength: 8
        type: string
      old_password:
        maxLength: 20
        minLength: 8
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  d7y_io_dragonfly_v2_manager_types.SignUpRequest:
    properties:
      avatar:
        type: string
      bio:
        type: string
      email:
        type: string
      location:
        type: string
      name:
        maxLength: 10
        minLength: 3
        type: string
      password:
        maxLength: 20
        minLength: 8
        type: string
      phone:
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  d7y_io_dragonfly_v2_manager_types.UpdateUserRequest:
    properties:
      avatar:
        type: string
      bio:
        type: string
      email:
        type: string
      location:
        type: string
      phone:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    url: https://d7y.io
  description: Dragonfly Manager Server
  license:
    name: Apache 2.0
  title: Dragonfly Manager
  version: 1.0.0
paths:
...
  /api/v1/user/signin/{name}:
    get:
      consumes:
      - application/json
      description: oauth signin by json config
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Oauth Signin
      tags:
      - User
  /api/v1/user/signin/{name}/callback:
    get:
      description: oauth signin callback by json config
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      - description: code
        in: query
        name: code
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Oauth Signin Callback
      tags:
      - Oauth
  /api/v1/user/signup:
    post:
      consumes:
      - application/json
      description: signup by json config
      parameters:
      - description: User
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SignUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.User'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: SignUp user
      tags:
      - User
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Get Users
      parameters:
      - default: 0
        description: current page
        in: query
        name: page
        required: true
        type: integer
      - default: 10
        description: return max item count, default 10, max 50
        in: query
        maximum: 50
        minimum: 2
        name: per_page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.User'
            type: array
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get Users
      tags:
      - User
  /api/v1/users/{id}:
    get:
      consumes:
      - application/json
      description: Get User by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.User'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get User
      tags:
      - User
    patch:
      consumes:
      - application/json
      description: Update by json config
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: User
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.User'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update User
      tags:
      - User
  /api/v1/users/{id}/reset_password:
    post:
      consumes:
      - application/json
      description: reset password by json config
      parameters:
      - description: User
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.ResetPasswordRequest'
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Reset Password For User
      tags:
      - User
  /api/v1/users/{id}/roles:
    get:
      description: get roles by json config
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get User Roles
      tags:
      - User
  /api/v1/users/{id}/roles/{role}:
    delete:
      consumes:
      - application/json
      description: delete role by uri config
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: role
        in: path
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Delete Role For User
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: add role to user by uri config
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: role
        in: path
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Add Role For User
      tags:
      - Users
...
swagger: "2.0"
...